src/requirements.txt
pandas==2.2.2
pytest==8.2.2
python-dotenv==1.0.1
PyYAML==6.0.1
Requests==2.32.3